scripts/requirements.txt
opencv-python
webvtt-py
